Updated --Directory structure
