+++ lib/services/auth_service.dart
@@ -1,0 +1,26 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+class AuthService {
+  final FirebaseAuth _auth = FirebaseAuth.instance;
+
+  // Stream to listen for auth state changes
+  Stream<User?> get authStateChanges => _auth.authStateChanges();
+
+  // Sign in anonymously
+  Future<void> signInAnonymously() async {
+    try {
+      await _auth.signInAnonymously();
+    } catch (e) {
+      throw Exception("Failed to sign in anonymously: $e");
+    }
+  }
+
+  // Sign out
+  Future<void> signOut() async {
+    try {
+      await _auth.signOut();
+    } catch (e) {
+      throw Exception("Failed to sign out: $e");
+    }
+  }
+}
